+++ public/index.html
@@ -4,6 +4,7 @@
         <meta charset="utf-8" />
         <meta name="viewport" content="width=device-width, initial-scale=1" />
         <title>Cozy House</title>
+        <link rel="icon" href="%PUBLIC_URL%/favicon.svg" />
     </head>
     <body>
         <div id="root"></div>
