+++ src/App.module.css
@@ -1,4 +1,4 @@
-.App {
+.app {
   text-align: center;
 }
