+++ src/components/DonationSection/DonationSection.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import classes from './DonationSection.module.css';
-import donationImage from '../../assets/images/donation-dog.webp';
+import donationImage from '../../assets/images/donationDog.webp';
 import creditCardImage from '../../assets/icons/creditCard.svg';
 
 const DonationSection = () => {
@@ -15,7 +15,7 @@ const DonationSection = () => {
                     <p className={classes.subHeader}>Name of the bank / Type of bank account</p>
                     <div className={classes.creditCard}>
                         <img src={creditCardImage} alt='Credit card'/>
-                        <h5>8380 2880 8028 8791 7435</h5>
+                        <p>8380 2880 8028 8791 7435</p>
                     </div>
                     <p className={classes.legalInfo}>Legal information and lorem ipsum dolor sit amet, consectetur adipiscing elit. Maecenas a ipsum at libero sagittis dignissim sed ac diam. Praesent ultrices maximus tortor et vulputate. Interdum et malesuada fames ac ante ipsum primis in faucibus.</p>
                 </div>
