+++ src/components/DonationSection/DonationSection.module.css
@@ -51,7 +51,7 @@
     border-radius: 9px;
 }
 
-.creditCard > h5 {
+.creditCard > p {
     font-family: Georgia;
     font-size: 20px;
     font-weight: 400;
