+++ src/components/HelpSection/HelpSection.module.css
@@ -67,7 +67,6 @@
     }
 }
 
-/* Instead of 320px, 470px is being used */
 @media (max-width: 470px) {
     .helpSectionContainer {
         padding: 42px 15px;
