+++ src/components/PetsSection/PetsSection.module.css
@@ -60,8 +60,8 @@
     }
     
     .header > h3 {
-        text-align: center;
         font-size: 25px;
         line-height: 32px;
+        text-align: center;
     }
 }
