+++ src/components/StartSection/StartSection.jsx
@@ -11,7 +11,7 @@ const StartSection = () => {
                 <div className={classes.contentContainer}>
                     <div className={classes.content}>
                         <h3>Not only people need a house</h3>
-                        <p>We offer to give a chance to a little and nice puppy with an extremely wide and open heart. He or she will love you more than anybody else in the world, you will see!</p>
+                        <h1>We offer to give a chance to a little and nice puppy with an extremely wide and open heart. He or she will love you more than anybody else in the world, you will see!</h1>
                         <button>Make a friend</button>
                     </div>
                     <img src={puppy} alt='Start screen puppy'/>
