+++ src/components/StartSection/StartSection.module.css
@@ -42,7 +42,7 @@
     color: #fff;
 }
 
-.content > p {
+.content > h1 {
     font-size: 15px;
     font-weight: 400;
     line-height: 24px;
@@ -107,7 +107,6 @@
     }
 }
 
-/* Instead of 320px, 470px is being used */
 @media (max-width: 470px) {
     .contentContainer {
         gap: 105px;
@@ -135,7 +134,7 @@
         text-align: center;
     }
 
-    .content > p {
+    .content > h1 {
         text-align: center;
         font-size: 15px;
     }
